requirements.txt
board==1.0
Pillow==9.0.1
python-dotenv==0.19.2
